+++ neural_net/src/layers.rs
@@ -3,12 +3,12 @@ use ndarray::Array2;
 use rand::Rng;
 use serde::{Deserialize, Serialize};
 
-use crate::activation::Activation;
+use crate::{activation::Activation, datasets::Float};
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct Layer {
-    pub weights: Array2<f64>,
-    pub biases: Array2<f64>,
+    pub weights: Array2<Float>,
+    pub biases: Array2<Float>,
     pub activation: Activation,
 }
 
@@ -20,7 +20,7 @@ pub struct GpuLayer {
 }
 
 impl Layer {
-    pub fn new(weights: Array2<f64>, biases: Array2<f64>, activation: Activation) -> Self {
+    pub fn new(weights: Array2<Float>, biases: Array2<Float>, activation: Activation) -> Self {
         Layer {
             weights,
             biases,
@@ -30,8 +30,8 @@ impl Layer {
 
     pub fn initialise_with(
         layer_sizes: (usize, usize),
-        weight_fn: fn(usize, usize) -> Array2<f64>,
-        bias_fn: fn(usize, usize) -> Array2<f64>,
+        weight_fn: fn(usize, usize) -> Array2<Float>,
+        bias_fn: fn(usize, usize) -> Array2<Float>,
         activation: Activation,
     ) -> Self {
         Layer {
@@ -87,27 +87,27 @@ pub fn default_leaky_relu(input_size: usize, output_size: usize) -> Layer {
     )
 }
 
-pub fn gen_random_matrix(input_size: usize, output_size: usize) -> Array2<f64> {
+pub fn gen_random_matrix(input_size: usize, output_size: usize) -> Array2<Float> {
     let mut rng = rand::rng();
     Array2::from_shape_fn((output_size, input_size), |_| rng.random_range(-1.0..1.0))
 }
 
-pub fn gen_he_matrix(input_size: usize, output_size: usize) -> Array2<f64> {
+pub fn gen_he_matrix(input_size: usize, output_size: usize) -> Array2<Float> {
     let mut rng = rand::rng();
-    let variance = (2.0 / (input_size as f64)).sqrt();
+    let variance = (2.0 / (input_size as Float)).sqrt();
     Array2::from_shape_fn((output_size, input_size), |_| {
         rng.random_range(-variance..variance)
     })
 }
 
-pub fn gen_xavier_matrix(input_size: usize, output_size: usize) -> Array2<f64> {
+pub fn gen_xavier_matrix(input_size: usize, output_size: usize) -> Array2<Float> {
     let mut rng = rand::rng();
-    let variance = (2.0 / (input_size + output_size) as f64).sqrt();
+    let variance = (2.0 / (input_size + output_size) as Float).sqrt();
     Array2::from_shape_fn((output_size, input_size), |_| {
         rng.random_range(-variance..variance)
     })
 }
 
-pub fn gen_zero_matrix(input_size: usize, output_size: usize) -> Array2<f64> {
+pub fn gen_zero_matrix(input_size: usize, output_size: usize) -> Array2<Float> {
     Array2::zeros((output_size, input_size))
 }
